pyproject.toml
[project]
name = "my-langchain-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13.9"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "ddgs>=9.10.0",
    "fastmcp>=0.9.0",
    "langchain>=0.2.6",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.6",
    "langchain-xai>=1.2.1",
    "python-dotenv>=1.2.1",
]
